MeDIC openEHR ResultSet translation refactored
